feat: Se realizó la sección de crear una contribución (85%). Además que se han agregado interfaces, tanto en backend como en el frontend. Y se agregaron mejoras del formulario con angular material.

+++ backend/src/main/java/com/revalisso/backend/entity/Estado.java
@@ -13,7 +13,7 @@ import java.util.List;
 public class Estado {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idEstado;
+    private Long idEstado;
     private String nombreEstado;
 
     @OneToMany(fetch = FetchType.LAZY, mappedBy = "estado", cascade = CascadeType.ALL)
